[download][s]: add click event listener to download buttons outside of react app.
This is useful/needed when your download button(s) are already in the page and outside of this react app.
So you can use `download-data` class with `value` attribute to specify format of data (json, csv, xlsx) and click event on that html element would trigger download action with the query in the state.